Fixed splitReflex() bug
Fixed a major bug where splitReflex() would crash the program for
various shapes.  The cause was a less than symbol mistaken for a greater
than symbol.  This caused a for loop that was meant to find the lowest
value from a list of vecs to find the highest, creating a cascade of
alternate problems and an eventual crash.  One more bug is left to be
fixed pertaining to the intersect() function in Utils.h.